+++ login-workflow/example3/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './DebugComponent';
+export * from './LanguageSelector';
