+++ login-workflow/example3/src/screens/ExampleHome.tsx
@@ -26,12 +26,11 @@ import Typography from '@mui/material/Typography';
 import { useTheme } from '@mui/material/styles';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import { FormControl, MenuItem, Select, SelectChangeEvent } from '@mui/material';
 import * as Colors from '@brightlayer-ui/colors';
+import { LanguageSelector } from '../components/LanguageSelector';
 
 export const ExampleHome: React.FC<React.PropsWithChildren> = () => {
     const app = useApp();
-    const [selectedLanguage, setSelectedLanguage] = useState(app.language);
     const { t } = useTranslation();
     const [open, setOpen] = useState(false);
     const navigate = useNavigate();
@@ -60,12 +59,6 @@ export const ExampleHome: React.FC<React.PropsWithChildren> = () => {
         justifyContent: 'center',
     };
 
-    const changeAppLanguage = (event: SelectChangeEvent): void => {
-        const appLanguage = event.target.value;
-        setSelectedLanguage(appLanguage);
-        app.setLanguage(appLanguage);
-    };
-
     const logOut = (): void => {
         LocalStorage.clearAuthCredentials();
         app.setIsAuthenticated(false);
@@ -92,32 +85,19 @@ export const ExampleHome: React.FC<React.PropsWithChildren> = () => {
                         <Typography variant="h6">{`${t('Blui:HOMEPAGE')}`}</Typography>
                         <Spacer />
                         <Box sx={{ py: 2 }}>
-                            <FormControl fullWidth>
-                                <Select
-                                    sx={{
-                                        padding: '4px',
-                                        minWidth: '160px',
-                                        margin: '0 16px',
-                                        backgroundColor: 'transparent',
+                            <LanguageSelector
+                                variant="standard"
+                                sx={{
+                                    padding: '4px',
+                                    minWidth: '160px',
+                                    margin: '16px',
+                                    backgroundColor: 'transparent',
+                                    color: Colors.white[50],
+                                    '& .MuiSelect-icon': {
                                         color: Colors.white[50],
-                                        '& .MuiSelect-icon': {
-                                            color: Colors.white[50],
-                                        },
-                                    }}
-                                    labelId="language-select-label"
-                                    id="language-select"
-                                    value={selectedLanguage}
-                                    label="Language"
-                                    variant="standard"
-                                    onChange={changeAppLanguage}
-                                >
-                                    <MenuItem value={'en'}>English</MenuItem>
-                                    <MenuItem value={'es'}>Spanish</MenuItem>
-                                    <MenuItem value={'fr'}>French</MenuItem>
-                                    <MenuItem value={'zh'}>Chinese</MenuItem>
-                                    <MenuItem value={'pt'}>Portugese</MenuItem>
-                                </Select>
-                            </FormControl>
+                                    },
+                                }}
+                            />
                         </Box>
                         <UserMenu
                             onClick={(): void => setOpen(!open)}
